new function to tag spiral scans, moved imaging and spiral scans to new sample location file
